refactor(frontend): split web/mobile app boundaries with target builds
